Move downloadFile to an external file
